refactor: split vsock guest connection tests
