+++ crates/vsock-guest/tests/connection/main.rs
@@ -1,0 +1,15 @@
+#![allow(
+    clippy::unwrap_used,
+    clippy::expect_used,
+    clippy::panic,
+    clippy::indexing_slicing
+)]
+
+mod support;
+
+mod exec_operation;
+mod process_control;
+mod quiesce;
+mod shutdown;
+mod spawn_buffered;
+mod spawn_streaming;
